Split the silence lanes out of lanes.go
